Fixed g++ include errors

+++ include/Graph.h
@@ -130,7 +130,7 @@ protected:
 
 private:
 	vector<Point> points;
-	Color lcolor {fl_color()};
+	Color lcolor {static_cast<int>(fl_color())};
 	Line_style ls {0};
 	Color fcolor {Color::invisible};
 };
